Share a loadConfig method between all java cloud client examples
