+++ clients/cloud/java/src/main/java/io/confluent/examples/clients/cloud/ConsumerExamplePageviews.java
@@ -15,21 +15,14 @@
  */
 package io.confluent.examples.clients.cloud;
 
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.Consumer;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
 import io.confluent.examples.clients.cloud.model.PageviewRecord;
 import io.confluent.kafka.serializers.KafkaJsonDeserializerConfig;
+import org.apache.kafka.clients.consumer.*;
 
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Properties;
+
+import static io.confluent.examples.clients.cloud.Util.loadConfig;
 
 public class ConsumerExamplePageviews {
 
@@ -71,17 +64,4 @@ public class ConsumerExamplePageviews {
       consumer.close();
     }
   }
-
-
-  public static Properties loadConfig(String configFile) throws IOException {
-    if (!Files.exists(Paths.get(configFile))) {
-      throw new IOException(configFile + " not found.");
-    }
-    final Properties cfg = new Properties();
-    try (InputStream inputStream = new FileInputStream(configFile)) {
-      cfg.load(inputStream);
-    }
-    return cfg;
-  }
-
 }
