+++ clients/cloud/java/src/main/java/io/confluent/examples/clients/cloud/StreamsExample.java
@@ -16,39 +16,28 @@
 
 package io.confluent.examples.clients.cloud;
 
-import java.util.Properties;
-
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-
-import org.apache.kafka.streams.KafkaStreams;
-import org.apache.kafka.streams.StreamsConfig;
-import org.apache.kafka.streams.kstream.KStream;
-import org.apache.kafka.streams.StreamsBuilder;
-import org.apache.kafka.streams.KeyValue;
-import org.apache.kafka.streams.kstream.KGroupedStream;
-import org.apache.kafka.streams.kstream.Printed;
-import org.apache.kafka.streams.kstream.Consumed;
-import org.apache.kafka.streams.kstream.Grouped;
-import org.apache.kafka.streams.kstream.Materialized;
-import org.apache.kafka.common.utils.Bytes;
-import org.apache.kafka.streams.state.KeyValueStore;
-
 import io.confluent.examples.clients.cloud.model.DataRecord;
+import io.confluent.kafka.serializers.KafkaJsonDeserializer;
+import io.confluent.kafka.serializers.KafkaJsonSerializer;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.common.serialization.Serializer;
-import io.confluent.kafka.serializers.KafkaJsonSerializer;
-import io.confluent.kafka.serializers.KafkaJsonDeserializer;
+import org.apache.kafka.streams.KafkaStreams;
+import org.apache.kafka.streams.KeyValue;
+import org.apache.kafka.streams.StreamsBuilder;
+import org.apache.kafka.streams.StreamsConfig;
+import org.apache.kafka.streams.kstream.Consumed;
+import org.apache.kafka.streams.kstream.Grouped;
+import org.apache.kafka.streams.kstream.KStream;
+import org.apache.kafka.streams.kstream.Printed;
 
-import java.util.Collections;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.io.FileInputStream;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Properties;
+
+import static io.confluent.examples.clients.cloud.Util.loadConfig;
 
 public class StreamsExample {
 
@@ -109,16 +98,4 @@ public class StreamsExample {
 
         return Serdes.serdeFrom(mySerializer, myDeserializer);
     }
-
-    public static Properties loadConfig(final String configFile) throws IOException {
-      if (!Files.exists(Paths.get(configFile))) {
-        throw new IOException(configFile + " not found.");
-      }
-      final Properties cfg = new Properties();
-      try (InputStream inputStream = new FileInputStream(configFile)) {
-        cfg.load(inputStream);
-      }
-      return cfg;
-    }
-
 }
